+++ src/app/lib/util.spec.ts
@@ -18,4 +18,20 @@ describe('Util', () => {
     reply = Util.traverseAncestors(b, (n) => {return true});
     expect(reply).toEqual([b, a]);
   });
+
+  it('should reject file names', () => {
+    expect(Util.validateFile(<File>null)).toBeNull();
+    expect(Util.validateFile(<File>{name: null})).toBeNull();
+    expect(Util.validateFile(<File>{name: '~a.a'})).toBeNull();
+    expect(Util.validateFile(<File>{name: '.a.a'})).toBeNull();
+    expect(Util.validateFile(<File>{name: '/a.a'})).toBeNull();
+    expect(Util.validateFile(<File>{name: 'a\\a'})).toBeNull();
+  });
+
+  it('should accept file names', () => {
+    expect(Util.validateFile(<File>{name: 'a.a'})).toBeDefined();
+    expect(Util.validateFile(<File>{name: 'aa'})).toBeDefined();
+    expect(Util.validateFile(<File>{name: 'a a'})).toBeDefined();
+    expect(Util.validateFile(<File>{name: 'A b.c'})).toBeDefined();
+  });
 });